fastapi/requirement.txt
transformers
pillow
tf-keras
tensorflow
fastapi
requests
uvicorn
torch
